+++ marker_adapter/src/lib.rs
@@ -12,6 +12,7 @@ pub use error::{Error, Result};
 pub use loader::LintCrateInfo;
 
 use loader::LintCrateRegistry;
+use marker_api::Lint;
 use marker_api::{
     ast::{Body, Crate, EnumVariant, ExprKind, ItemField, ItemKind, StmtKind},
     context::MarkerContext,
@@ -54,8 +55,16 @@ impl Adapter {
         })
     }
 
+    pub fn marker_lints(&self) -> Vec<&'static Lint> {
+        self.lint_pass_infos()
+            .iter()
+            .flat_map(marker_api::LintPassInfo::lints)
+            .copied()
+            .collect()
+    }
+
     #[must_use]
-    pub fn lint_pass_infos(&self) -> Vec<LintPassInfo> {
+    fn lint_pass_infos(&self) -> Vec<LintPassInfo> {
         self.inner.borrow().external_lint_crates.collect_lint_pass_info()
     }