+++ marker_lints/src/lib.rs
@@ -1,17 +1,10 @@
 #![doc = include_str!("../README.md")]
 #![feature(let_chains)]
 
-use marker_api::{prelude::*, LintPass, LintPassInfo, LintPassInfoBuilder};
+mod diag_msg_uppercase_start;
+mod not_using_has_span_trait;
 
-marker_api::declare_lint! {
-    /// ### What it does
-    /// Diagnostic messages should start with lower case letter according to
-    /// [rustc's dev guide].
-    ///
-    /// [rustc's dev guide]: <https://rustc-dev-guide.rust-lang.org/diagnostics.html#diagnostic-output-style-guide>
-    DIAG_MSG_UPPERCASE_START,
-    Warn,
-}
+use marker_api::{prelude::*, LintPass, LintPassInfo, LintPassInfoBuilder};
 
 #[derive(Debug, Default)]
 struct MarkerLintsLintPass;
@@ -20,32 +13,19 @@ marker_api::export_lint_pass!(MarkerLintsLintPass);
 
 impl LintPass for MarkerLintsLintPass {
     fn info(&self) -> LintPassInfo {
-        LintPassInfoBuilder::new(Box::new([DIAG_MSG_UPPERCASE_START])).build()
+        let lints = [
+            diag_msg_uppercase_start::DIAG_MSG_UPPERCASE_START,
+            not_using_has_span_trait::NOT_USING_HAS_SPAN_TRAIT,
+        ];
+
+        LintPassInfoBuilder::new(Box::new(lints)).build()
     }
 
     fn check_expr<'ast>(&mut self, cx: &MarkerContext<'ast>, expr: ExprKind<'ast>) {
-        if let ExprKind::Method(call) = expr
-            && let sem::TyKind::Ref(adt_ref) = call.receiver().ty()
-            && let sem::TyKind::Adt(adt) = adt_ref.inner_ty()
-            && cx.resolve_ty_ids("marker_api::MarkerContext").contains(&adt.def_id())
-            && call.method().ident().name() == "emit_lint"
-            && let [_lint, _id, msg, ..] = call.args()
-        {
-            check_msg(cx, *msg);
-        }
+        diag_msg_uppercase_start::check_expr(cx, expr);
     }
-}
 
-fn check_msg<'ast>(cx: &MarkerContext<'ast>, msg_expr: ExprKind<'ast>) {
-    if let ExprKind::StrLit(lit) = msg_expr
-        && let Some(msg) = lit.str_value()
-        && let Some(start) = msg.chars().next()
-        && start.is_uppercase()
-    {
-        cx.emit_lint(
-            DIAG_MSG_UPPERCASE_START,
-            msg_expr,
-            "this message starts with an uppercase character",
-        );
+    fn check_item<'ast>(&mut self, cx: &'ast MarkerContext<'ast>, item: ast::ItemKind<'ast>) {
+        not_using_has_span_trait::check_item(cx, item);
     }
 }